requirements-autonomous.txt
pandas==2.0.1
numpy==1.23.5
scikit-learn==1.2.2
joblib==1.2.0
requests==2.28.2
google-cloud-firestore==2.9.1
python-dateutil==2.8.2